Apply tab pane for types
Split up tab pane factory in two, since they have nothing in common,
only require what’s necessary.

Placeholders for the individual types. Styles are obviously wrong, TBD
later.
