+++ lib/assets/javascripts/cartodb3/editor.js
@@ -7,7 +7,7 @@ var EditorMapView = require('./editor/map/map-view');
 var LayerDefinitionsCollection = require('./data/layer-definitions-collection');
 var WidgetDefinitionsCollection = require('./data/widget-definitions-collection');
 var VisDefinitionModel = require('./data/vis-definition-model');
-var TabPaneViewFactory = require('./components/tab-pane/tab-pane-factory');
+var createEditorMenuTabPane = require('./components/tab-pane/create-editor-menu-tab-pane');
 var EditorPaneTemplate = require('./editor-pane.tpl');
 var ModalsServiceModel = require('./components/modals/modals-service-model');
 var viewFactory = require('./components/view-factory');
@@ -77,7 +77,7 @@ dashboard.vis.done(function () {
 
 var modals = new ModalsServiceModel();
 
-var editorTabPaneView = TabPaneViewFactory.createWithIcons([
+var editorTabPaneView = createEditorMenuTabPane([
   {
     icon: 'dashboard',
     createContentView: function () {
