+++ lib/assets/javascripts/cartodb3/editor/add-widgets/add-widgets-body-view.js
@@ -1,5 +1,8 @@
 var cdb = require('cartodb.js');
 var AddWidgetsFormulaOptionsView = require('./formula/add-widgets-formula-options-view');
+var createTextLabelsTabPane = require('../../components/tab-pane/create-text-labels-tab-pane');
+var tabPaneTemplate = require('./add-widgets-tab-pane-template.tpl');
+var viewFactory = require('../../components/view-factory');
 
 /**
  * View to select widget options to create.
@@ -47,14 +50,46 @@ module.exports = cdb.core.View.extend({
   render: function () {
     this.clearSubViews();
 
-    // TODO should be wrapped in a tabpane component here
-    var view = new AddWidgetsFormulaOptionsView({
-      className: 'Dialog-bodyInnerExpandedWithSubFooter',
-      buckets: this._buckets,
-      selectedCollection: this._selectedCollection
-    });
+    var buckets = this._buckets;
+    var selectedCollection = this._selectedCollection;
+
+    var tabPaneItems = [{
+      label: _t('editor.add-widgets.tab-pane.histogram-label'),
+      createContentView: viewFactory.createByHTML.bind(viewFactory, 'TBD')
+    }, {
+      label: _t('editor.add-widgets.tab-pane.category-label'),
+      createContentView: viewFactory.createByHTML.bind(viewFactory, 'TBD')
+    }, {
+      selected: true,
+      label: _t('editor.add-widgets.tab-pane.formula-label'),
+      createContentView: function () {
+        return new AddWidgetsFormulaOptionsView({
+          buckets: buckets,
+          selectedCollection: selectedCollection
+        });
+      }
+    }, {
+      label: _t('editor.add-widgets.tab-pane.time-series-label'),
+      createContentView: viewFactory.createByHTML.bind(viewFactory, 'TBD')
+    }];
+    var options = {
+      tabPaneOptions: {
+        tagName: 'nav',
+        className: 'CDB-NavMenu',
+        template: tabPaneTemplate,
+        tabPaneItemOptions: {
+          tagName: 'li',
+          className: 'CDB-NavMenu-Item'
+        }
+      },
+      tabPaneItemLabelOptions: {
+        tagName: 'button',
+        className: 'CDB-NavMenu-Link'
+      }
+    };
+    var view = createTextLabelsTabPane(tabPaneItems, options);
     this.addView(view);
-    this.$el.append(view.render().$el);
+    this.$el.append(view.render().el);
 
     return this;
   }
