+++ lib/assets/javascripts/cartodb3/editor/map/map-view.js
@@ -2,7 +2,7 @@ var $ = require('jquery');
 var Backbone = require('backbone');
 var cdb = require('cartodb-deep-insights.js');
 var EditorWidgetsView = require('../widgets/widgets-view');
-var TabPaneViewFactory = require('../../components/tab-pane/tab-pane-factory');
+var createTextLabelsTabPane = require('../../components/tab-pane/create-text-labels-tab-pane');
 var StackLayoutView = require('../../components/stack-layout/stack-layout-view');
 var WidgetsFormContentView = require('../widgets/widgets-form/widgets-form-content-view');
 var Template = require('./template.tpl');
@@ -49,7 +49,7 @@ module.exports = cdb.core.View.extend({
       }
     }]);
 
-    var mapTabPaneView = TabPaneViewFactory.createWithTextLabels([{
+    var mapTabPaneView = createTextLabelsTabPane([{
       label: _t('editor.layers.title'),
       selected: true,
       createContentView: function () {
