+++ lib/assets/javascripts/cartodb3/editor/widgets/widgets-form/widgets-form-content-view.js
@@ -2,7 +2,7 @@ var cdb = require('cartodb-deep-insights.js');
 var $ = require('jquery');
 var WidgetsFormContentDataView = require('./widgets-form-content-data-view');
 var WidgetsFormContentStyleView = require('./widgets-form-content-style-view');
-var TabPaneViewFactory = require('../../../components/tab-pane/tab-pane-factory');
+var createTextLabelsTabPane = require('../../../components/tab-pane/create-text-labels-tab-pane');
 var TabPaneTemplate = require('../../map/tab-pane-template.tpl');
 
 /**
@@ -77,7 +77,7 @@ module.exports = cdb.core.View.extend({
       }
     };
 
-    this.tabPaneView = TabPaneViewFactory.createWithTextLabels(this.tabPaneItems, options);
+    this.tabPaneView = createTextLabelsTabPane(this.tabPaneItems, options);
     this.$el.append(this.tabPaneView.render().$el);
   },
