+++ lib/assets/test/spec/cartodb3/components/tab-pane/create-editor-menu-tab-pane.spec.js
@@ -1,0 +1,71 @@
+var _ = require('underscore');
+var cdb = require('cartodb-deep-insights.js');
+var createEditorMenuTabPane = require('../../../../../javascripts/cartodb3/components/tab-pane/create-editor-menu-tab-pane');
+
+describe('components/tab-pane/create-editor-menu-tab-pane', function () {
+  beforeEach(function () {
+    this.items = [
+      {
+        selected: false,
+        icon: 'myFirstIcon',
+        createContentView: function () {
+          return new cdb.core.View();
+        }
+      }, {
+        selected: false,
+        icon: 'mySecondIcon',
+        createContentView: function () {
+          return new cdb.core.View();
+        }
+      }
+    ];
+  });
+
+  it('should create view', function () {
+    var view = createEditorMenuTabPane(this.items);
+    expect(view).toEqual(jasmine.any(cdb.core.View));
+    expect(view.collection.size()).toEqual(2);
+
+    view.render();
+    expect(_.size(view._subviews)).toEqual(3);
+  });
+
+  it('should add the icons', function () {
+    var view = createEditorMenuTabPane(this.items);
+    view.render();
+    expect(view.$el.find('i.CDB-IconFont-myFirstIcon').length).toBe(1);
+    expect(view.$el.find('i.CDB-IconFont-mySecondIcon').length).toBe(1);
+  });
+
+  it('should throw error if icon is not provided', function () {
+    var items = _.clone(this.items);
+    delete items[0].icon;
+    expect(function () { createEditorMenuTabPane(items); }).toThrow(new Error('icon should be provided'));
+  });
+
+  it('should throw error if contentView is not provided', function () {
+    var items = _.clone(this.items);
+    delete items[0].createContentView;
+    expect(function () { createEditorMenuTabPane(items); }).toThrow(new Error('createContentView should be provided'));
+  });
+
+  it('should allow to set custom options for the pane and its items', function () {
+    var view = createEditorMenuTabPane(this.items, {
+      tabPaneOptions: {
+        className: 'MyCoolIconTabPane',
+        tabPaneItemOptions: {
+          tagName: 'li',
+          className: 'CDB-NavMenu-Item'
+        }
+      },
+      tabPaneItemIconOptions: {
+        tagName: 'button',
+        className: 'CDB-NavMenu-Link'
+      }
+    });
+    view.render();
+    expect(view.$el.hasClass('MyCoolIconTabPane')).toBeTruthy();
+    expect(view.$el.find('li.CDB-NavMenu-Item').length).toBe(2);
+    expect(view.$el.find('button.CDB-NavMenu-Link').length).toBe(2);
+  });
+});
